Organise code into include and source.
Running the `build.sh` will compile the code for this lecture series.

In theory this should be fairly straightforward to maintain for a small project like this. Although, might be useful to make a `.sh` script to automate the build process later. Right now, I'm just not going to need it.