Add demo for pipe communication.
Create a basic demo for pipe communication between two processes. This
demo creates two one-way pipes and a message is sent from each process
to the other, only to be printed.

Also replaced the `build.sh` script with a `Makefile` to make
compilation faster. Not exactly necessary for such a small project, but
it's a good habit to get into ig.

TODO: Add a demo for shared memory.